Added candidate state and district dynamic validation rules to schema
