+++ fecfile_validate_python/tests/sample_contact_IND.json
@@ -15,7 +15,7 @@
     "zip": "22204",
     "employer": "Byron Inc",
     "occupation": "Business Owner",
-    "candidate_office": "P",
+    "candidate_office": null,
     "candidate_state": null,
     "candidate_district": null,
     "telephone": "3043892120",
