+++ schema/Contact_Candidate.json
@@ -15,8 +15,6 @@
         "state",
         "zip",
         "candidate_office",
-        "candidate_state",
-        "candidate_district",
         "country"
     ],
     "fec_recommended": [],
@@ -222,9 +220,9 @@
             "title": "STATE",
             "description": "",
             "type": "string",
-            "minLength": 0,
+            "minLength": 2,
             "maxLength": 2,
-            "pattern": "^[ A-Za-z0-9]{0,2}$",
+            "pattern": "^[A-Z]{2}$",
             "examples": ["WA"],
             "fec_spec": {
                 "COL_SEQ": 11,
@@ -319,10 +317,10 @@
         "candidate_state": {
             "title": "CANDIDATE STATE",
             "description": "",
-            "type": "string",
-            "minLength": 0,
+            "type": ["string", "null"],
+            "minLength": 2,
             "maxLength": 2,
-            "pattern": "^[ A-Za-z0-9]{0,2}$",
+            "pattern": "^[A-Z]{2}$",
             "examples": ["WA"],
             "fec_spec": {
                 "COL_SEQ": 16,
@@ -339,10 +337,10 @@
         "candidate_district": {
             "title": "CANDIDATE DISTRICT",
             "description": "",
-            "type": "string",
-            "minLength": 0,
+            "type": ["string", "null"],
+            "minLength": 2,
             "maxLength": 2,
-            "pattern": "^[ A-Za-z0-9]{0,2}$",
+            "pattern": "^[0-9]{2}$",
             "fec_spec": {
                 "COL_SEQ": 17,
                 "FIELD_DESCRIPTION": "CANDIDATE DISTRICT",
@@ -390,5 +388,34 @@
                 "FIELD_FORM_ASSOCIATION": null
             }
         }
-    }
+    },
+    "allOf": [
+        {
+            "if": {
+                "properties": {
+                    "candidate_office": { "const": "S" }
+                },
+                "required": ["candidate_office"]
+            },
+            "then": {
+                "required": ["candidate_state"],
+                "properties": { "candidate_state": { "type": "string" } }
+            }
+        },
+        {
+            "if": {
+                "properties": {
+                    "candidate_office": { "const": "H" }
+                },
+                "required": ["candidate_office"]
+            },
+            "then": {
+                "required": ["candidate_state", "candidate_district"],
+                "properties": {
+                    "candidate_state": { "type": "string" },
+                    "candidate_district": { "type": "string" }
+                }
+            }
+        }
+    ]
 }
